requirements.txt
pandas==1.2.4
tensorflow-gpu==2.3.0
tqdm==4.61.0
sklearn==0.0